Remove builder, Docker tasker and obsolete prepub_tests_for_image plugin
* CLOUDBLD-6299
